Place light nodes precisely, without vector.floor()
This avoids flickering flashlights when the player stands very close to a wall.
Flickering happened probably because floor() moved the position to a node corner,
and then minetest.get_node() and minetest.set_node() operated on an adjacent node,
which is not lightable because it is part of the wall.